GKMap: refactoring

+++ GKMap.WinForms/GMapRoundedToolTip.cs
@@ -47,7 +47,7 @@ namespace GKMap.WinForms
         {
             Size st = g.MeasureString(Marker.ToolTipText, Font).ToSize();
 
-            Rectangle rect = new Rectangle(Marker.ToolTipPosition.X, Marker.ToolTipPosition.Y - st.Height, st.Width + TextPadding.Width * 2, st.Height + TextPadding.Height);
+            Rectangle rect = new Rectangle((int)Marker.ToolTipPosition.X, (int)Marker.ToolTipPosition.Y - st.Height, st.Width + TextPadding.Width * 2, st.Height + TextPadding.Height);
             rect.Offset(Offset.X, Offset.Y);
 
             g.DrawLine(Stroke, Marker.ToolTipPosition.X, Marker.ToolTipPosition.Y, rect.X + Radius / 2, rect.Y + rect.Height - Radius / 2);
